% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogo.R
\name{catalogo_ece}
\alias{catalogo_ece}
\title{Consulta el catálogo de la Encuesta Continua de Empleo (ECE)}
\usage{
catalogo_ece(anio = NULL, trimestre = NULL, tabla = NULL)
}
\arguments{
\item{anio}{Entero opcional. Filtra por año.}

\item{trimestre}{Entero opcional (1-4). Filtra por trimestre.}

\item{tabla}{Caracteres opcional. Filtra por nivel.}
}
\value{
Un data.frame con las filas del catálogo correspondientes a la ECE.
}
\description{
Consulta el catálogo de la Encuesta Continua de Empleo (ECE)
}
\examples{
catalogo_ece()
catalogo_ece(anio = 2022, trimestre = 3)
}
