% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogo.R
\name{catalogo_eh}
\alias{catalogo_eh}
\title{Consulta el catálogo de la Encuesta de Hogares (EH)}
\usage{
catalogo_eh(anio = NULL, tabla = NULL)
}
\arguments{
\item{anio}{Entero opcional. Filtra por año.}

\item{tabla}{Caracteres opcional. Filtra por nivel (e.g., `"persona"`).}
}
\value{
Un data.frame con las filas del catálogo correspondientes a la EH.
}
\description{
Consulta el catálogo de la Encuesta de Hogares (EH)
}
\examples{
catalogo_eh()
catalogo_eh(anio = 2023)
}
