% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{departamentos}
\alias{departamentos}
\title{Lista los departamentos de Bolivia}
\usage{
departamentos()
}
\value{
Un data.frame con columnas `depto` (código entero 1-9) y `nombre_dep`.
}
\description{
Lista los departamentos de Bolivia
}
\examples{
departamentos()
}
