% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encuestasbo-package.R
\docType{package}
\name{encuestasbo-package}
\alias{encuestasbo}
\alias{encuestasbo-package}
\title{encuestasbo: Acceso, armonización y análisis con diseño muestral de las encuestas de hogares y empleo del INE de Bolivia (2012-2024)}
\description{
Proporciona acceso programático a los microdatos de las encuestas del Instituto Nacional de Estadística (INE) de Bolivia: la Encuesta de Hogares (EH, 2012-2024) y la Encuesta Continua de Empleo (ECE, 2015+). Descarga archivos Parquet desde GitHub Releases, los almacena en caché local y permite filtrar por departamento y área. Soporta flujos de trabajo estilo dplyr vía Apache Arrow y consultas SQL vía DuckDB. Incluye diccionarios de variables por encuesta y año, una capa de armonización canónica entre años, indicadores derivados (empleo, educación, pobreza, ingresos) y helpers de diseño muestral con 'srvyr'/'survey' para producir estimaciones e intervalos de confianza estadísticamente correctos.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lab-tecnosocial.github.io/encuestasbo/}
  \item \url{https://github.com/lab-tecnosocial/encuestasbo}
  \item Report bugs at \url{https://github.com/lab-tecnosocial/encuestasbo/issues}
}

}
\author{
\strong{Maintainer}: Alex Ojeda Copa \email{alex@labtecnosocial.org} (organization: Lab TecnoSocial)

Authors:
\itemize{
  \item Alex Ojeda Copa \email{alex@labtecnosocial.org} (organization: Lab TecnoSocial)
}

}
\keyword{internal}
