% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armonizar.R
\name{grupos_variables}
\alias{grupos_variables}
\title{Grupos temáticos de variables armonizadas}
\usage{
grupos_variables()
}
\value{
Una lista nombrada de grupos con los nombres canónicos de variables.
}
\description{
Grupos temáticos de variables armonizadas
}
\examples{
grupos_variables()
grupos_variables()$pobreza
}
