% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armonizar.R
\name{variables_armonizadas}
\alias{variables_armonizadas}
\title{Lista las variables armonizadas disponibles}
\usage{
variables_armonizadas(solo_armonizadas = TRUE)
}
\arguments{
\item{solo_armonizadas}{Lógico. Si `TRUE` (defecto), solo las marcadas como
comparables entre años.}
}
\value{
Un data.frame con las columnas del [variable_canonica_map].
}
\description{
Lista las variables armonizadas disponibles
}
\examples{
variables_armonizadas()
}
