% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armonizar_ece.R
\name{variables_armonizadas_ece}
\alias{variables_armonizadas_ece}
\title{Lista las columnas canónicas de la serie ECE armonizada}
\usage{
variables_armonizadas_ece()
}
\value{
Un data.frame con el nombre canónico y una breve descripción de cada
  columna que devuelve [get_ece_armonizada()].
}
\description{
Lista las columnas canónicas de la serie ECE armonizada
}
\examples{
variables_armonizadas_ece()
}
\seealso{
[get_ece_armonizada()], [variables_armonizadas()].
}
